(CURRENT) OnlyFunds/requirements.txt
# Core
pandas>=1.3
numpy>=1.21
requests>=2.25
python-dotenv>=1.0
ta>=0.10  # Technical Analysis indicators
scikit-learn>=1.0

# Web App & UI
streamlit>=1.25
plotly>=5.0
matplotlib>=3.4

# Backtesting
backtesting>=0.3
vectorbt>=0.25  # Optional, very powerful

# Hyperparameter tuning (AI AutoTune)
optuna>=3.1
scipy>=1.7

# Extras
tqdm>=4.62
rich>=13.0

# Notifications
python-telegram-bot>=13.15

# If running on Python < 3.10
typing-extensions>=4.0
